Render the same TaskRow on the dashboard and in the terminal
Follow-up to the unified toolbar: the rows themselves still differed.
The terminal showed a square checkbox + an amber "highest-priority"
star; the dashboard showed a round toggle, no star, and a lighter
layout. Zack: "you didn't add the check boxes or the priority stars in
the dashboard task list like they're shown in the terminal... make sure
they're the same."

Make the row a single shared component so the two surfaces can't drift:

- Extract TaskRow (+ InlineTitleEditor + recurrence helpers) out of
  TasksPane into components/TaskRow.tsx. Terminal-only affordances
  (drag handle, subtask expander, comments, request-update, inline
  rename, row selection) are now driven by OPTIONAL handlers — omit a
  handler and its control isn't rendered. Added an optional
  `terminalName` chip (the dashboard list spans terminals) and a narrow
  `TaskRowTask` type both the terminal Task and the dashboard
  AssignedTask satisfy.
- TasksPane imports the shared TaskRow and passes every handler, so the
  in-terminal pane is byte-for-byte unchanged. Removed the now-dead
  local copies and orphaned imports.
- dashboard-queries: select `starred` in loadAssignedTasks +
  loadDelegatedTasks and add it to AssignedTask, so dashboard rows can
  render the star.
- TasksCard: replace the bespoke AssignedRow with a thin
  DashboardTaskRow wrapper that owns optimistic done/star state and
  delegates rendering to the shared TaskRow. Toggling done or star
  PATCHes the same /api/v1/tasks/:id endpoint the terminal uses.

Result: dashboard rows now show the identical square checkbox, amber
star, priority left-edge + dots, due chip and status pill as the
terminal — one component, one source of truth.

Typecheck + lint clean; 315/315 web unit tests pass.
