+++ apps/web/src/components/TasksPane.tsx
@@ -2,20 +2,7 @@
 
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import dynamic from "next/dynamic";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
-import {
-  ChevronDown,
-  ChevronRight,
-  Plus,
-  Check,
-  MessageSquare,
-  Maximize2,
-  ListTodo,
-  Repeat,
-  Send,
-  Star,
-} from "lucide-react";
+import { Plus, Check, MessageSquare, ListTodo } from "lucide-react";
 import { cn } from "@/lib/utils";
 import { EmptyState } from "./EmptyState";
 import { useRealtimeTable } from "@/lib/supabase/realtime";
@@ -34,17 +21,9 @@ import {
   type TaskComposerSubmit,
 } from "./TaskComposer";
 import { SubtasksList, type Subtask } from "./SubtasksList";
-import {
-  PriorityDots,
-  StatusPill,
-  DueChip,
-} from "./primitives";
-import {
-  TaskSectionHeader,
-  groupTone,
-  priorityEdge,
-} from "./TaskSectionHeader";
+import { TaskSectionHeader, groupTone } from "./TaskSectionHeader";
 import { TaskListToolbar, type GroupOption } from "./TaskListToolbar";
+import { TaskRow } from "./TaskRow";
 import { groupTasks, type TaskGroupMode } from "@/lib/task-grouping";
 
 /** Group-by options for the in-terminal pane (no "Terminal" — you're
@@ -1070,399 +1049,6 @@ export function TasksPane({ ticker, projectId, currentUserId }: TasksPaneProps) 
         </div>
       ) : null}
     </div>
-  );
-}
-
-function TaskRow({
-  task,
-  ticker,
-  selected,
-  expanded,
-  draggable,
-  onClick,
-  onToggle,
-  onToggleStar,
-  onOpenComments,
-  onToggleExpand,
-  onRequestUpdate,
-  onRename,
-}: {
-  task: Task;
-  ticker: string;
-  selected: boolean;
-  expanded: boolean;
-  draggable: boolean;
-  onClick: () => void;
-  onToggle: () => void;
-  /**
-   * Flip the "highest priority of the day" star. Starred rows float
-   * to the top of the list regardless of priority/due/position.
-   */
-  onToggleStar: () => void;
-  onOpenComments: () => void;
-  onToggleExpand: () => void;
-  onRequestUpdate: () => void;
-  /**
-   * Persist a renamed title. Called when the user blurs / hits
-   * Enter on the inline-edit input. Receives the trimmed new title;
-   * implementation handles validation + PATCH + optimistic rollback.
-   */
-  onRename: (nextTitle: string) => void;
-}) {
-  const done = task.status === "done";
-  const subtaskTotal = task.subtask_total ?? 0;
-  const subtaskDone = task.subtask_done ?? 0;
-  const status = task.latest_status_text?.trim() ?? "";
-  const externalCount = task.external_assignee_emails?.length ?? 0;
-  const externalEmailsTitle =
-    externalCount > 0
-      ? `External assignees: ${task.external_assignee_emails!.join(", ")}`
-      : "";
-
-  return (
-    <div
-      onClick={onClick}
-      className={cn(
-        // Always reserve a 2px left edge (pl-[6px] + 2px border = the
-        // base px-2) so colouring it never shifts the row. The edge
-        // signals priority at a glance — High red, Medium amber — and
-        // is overridden by the amber selection edge when the row is
-        // active.
-        "group flex cursor-pointer items-center gap-2 border-l-2 px-2 py-2.5 pl-[6px] transition-colors",
-        selected ? "bg-bg-2" : "hover:bg-bg-2",
-        selected ? "border-l-border-focus" : priorityEdge(task.priority),
-      )}
-    >
-      {/* Drag handle. Only rendered when the parent is in Manual sort
-          mode (the actual draggable attribute is on the <li>); the
-          handle is a hint that the row CAN be dragged. */}
-      {draggable ? (
-        <span
-          aria-hidden="true"
-          title="Drag to reorder"
-          className="flex h-4 w-3 flex-shrink-0 cursor-grab items-center justify-center text-text-3 group-hover:text-text-1 active:cursor-grabbing"
-        >
-          <span className="grid h-3 w-1.5 grid-cols-2 grid-rows-3 gap-[1px]">
-            {[0, 1, 2, 3, 4, 5].map((n) => (
-              <span
-                key={n}
-                className="block h-[2px] w-[2px] rounded-full bg-current"
-              />
-            ))}
-          </span>
-        </span>
-      ) : null}
-      <button
-        type="button"
-        onClick={(e) => {
-          e.stopPropagation();
-          onToggleExpand();
-        }}
-        aria-label={expanded ? "Collapse subtasks" : "Expand subtasks"}
-        aria-expanded={expanded}
-        className="flex h-4 w-4 flex-shrink-0 items-center justify-center rounded-sm text-text-3 hover:bg-bg-3 hover:text-text-0"
-      >
-        {expanded ? (
-          <ChevronDown className="h-3 w-3" aria-hidden="true" />
-        ) : (
-          <ChevronRight className="h-3 w-3" aria-hidden="true" />
-        )}
-      </button>
-      <button
-        type="button"
-        onClick={(e) => {
-          e.stopPropagation();
-          onToggleStar();
-        }}
-        aria-label={task.starred ? "Unstar (remove from top)" : "Star (pin to top)"}
-        aria-pressed={task.starred}
-        title={
-          task.starred
-            ? "Starred — highest priority. Click to unstar."
-            : "Star to pin to top of the list"
-        }
-        className={cn(
-          "flex h-4 w-4 flex-shrink-0 items-center justify-center rounded-sm transition-colors",
-          task.starred
-            ? "text-warning"
-            : "text-text-3 hover:text-warning opacity-0 group-hover:opacity-100",
-          // Starred state stays visible always — even when not hovered —
-          // because the star is the highest-priority signal on the row.
-        )}
-      >
-        <Star
-          className="h-3 w-3"
-          fill={task.starred ? "currentColor" : "none"}
-          aria-hidden="true"
-        />
-      </button>
-      <button
-        onClick={(e) => {
-          e.stopPropagation();
-          onToggle();
-        }}
-        aria-label={done ? "Mark as not done" : "Mark as done"}
-        className={cn(
-          "flex h-4 w-4 flex-shrink-0 items-center justify-center rounded-sm border",
-          done
-            ? "border-success bg-success-subtle text-success"
-            : "border-border hover:border-accent",
-        )}
-      >
-        {done ? <Check className="h-3 w-3" aria-hidden="true" /> : null}
-      </button>
-      <div className="flex min-w-0 flex-1 flex-col gap-0.5">
-        <InlineTitleEditor
-          title={task.title}
-          done={done}
-          href={`/p/${ticker}/task/${task.ticker_seq}`}
-          onCommit={onRename}
-        />
-        {status ? (
-          <span
-            className="flex items-center gap-1 truncate text-[11px] leading-tight text-text-2"
-            title={status}
-          >
-            <span className="font-mono text-[9px] uppercase tracking-wide text-text-3">
-              Status
-            </span>
-            <span className="truncate">{status}</span>
-          </span>
-        ) : null}
-      </div>
-      {/* Subtask roll-up — surfaces the count without expanding. The
-          list endpoint already returns subtask_total/subtask_done
-          aggregates, so this is free. */}
-      {subtaskTotal > 0 ? (
-        <span
-          className="flex-shrink-0 rounded-sm bg-bg-3 px-1 font-mono text-[10px] text-text-2"
-          title={`${subtaskDone} of ${subtaskTotal} subtasks done`}
-        >
-          {subtaskDone}/{subtaskTotal}
-        </span>
-      ) : null}
-      {externalCount > 0 ? (
-        <span
-          className="flex-shrink-0 rounded-sm border border-border bg-bg-2 px-1 font-mono text-[10px] uppercase tracking-wide text-text-2"
-          title={externalEmailsTitle}
-        >
-          @+{externalCount}
-        </span>
-      ) : null}
-      <Link
-        href={`/p/${ticker}/task/${task.ticker_seq}`}
-        onClick={(e) => e.stopPropagation()}
-        aria-label="Open task detail"
-        className="rounded-sm p-1 text-text-3 opacity-0 transition-opacity hover:bg-bg-3 hover:text-text-0 group-hover:opacity-100"
-      >
-        <Maximize2 className="h-3 w-3" />
-      </Link>
-      <button
-        onClick={(e) => {
-          e.stopPropagation();
-          onOpenComments();
-        }}
-        aria-label="Comments"
-        className="rounded-sm p-1 text-text-3 opacity-0 transition-opacity hover:bg-bg-3 hover:text-text-0 group-hover:opacity-100"
-      >
-        <MessageSquare className="h-3 w-3" />
-      </button>
-      <button
-        onClick={(e) => {
-          e.stopPropagation();
-          onRequestUpdate();
-        }}
-        aria-label="Request status update"
-        title="Request status update"
-        className="rounded-sm p-1 text-text-3 opacity-0 transition-opacity hover:bg-bg-3 hover:text-accent group-hover:opacity-100"
-      >
-        <Send className="h-3 w-3" />
-      </button>
-      {task.due_date ? <DueChip date={task.due_date} /> : null}
-      {task.recurrence_rule ? (
-        <span
-          className="flex flex-shrink-0 items-center gap-0.5 rounded-sm border border-border bg-bg-2 px-1 py-0.5 text-[10px] uppercase tracking-wide text-text-2"
-          title={`Repeats ${recurrenceLabel(task.recurrence_rule)}`}
-        >
-          <Repeat className="h-2.5 w-2.5" aria-hidden="true" />
-          {recurrenceShortLabel(task.recurrence_rule)}
-        </span>
-      ) : null}
-      <PriorityDots priority={task.priority} />
-      <StatusPill status={task.status} />
-    </div>
-  );
-}
-
-/** Long form for tooltips: "Daily", "Weekly", "Monthly ×2". */
-function recurrenceLabel(rule: TaskRecurrenceRule): string {
-  const base =
-    rule.pattern === "daily"
-      ? "Daily"
-      : rule.pattern === "weekly"
-        ? "Weekly"
-        : "Monthly";
-  return rule.interval > 1 ? `${base} ×${rule.interval}` : base;
-}
-
-/** Single-char chip glyph: "D", "W", "M" (+ optional interval). */
-function recurrenceShortLabel(rule: TaskRecurrenceRule): string {
-  const letter =
-    rule.pattern === "daily" ? "D" : rule.pattern === "weekly" ? "W" : "M";
-  return rule.interval > 1 ? `${letter}${rule.interval}` : letter;
-}
-
-/**
- * Inline title editor for a task row. Renders as a clickable title in
- * its default state — single click navigates to the task detail page,
- * double-click flips to a focused `<input>` for rename. Enter commits,
- * Escape cancels, blur commits the current value.
- *
- * Single vs. double click is disambiguated with a short timer: the
- * navigate fires ~220ms after click unless a dblclick lands first and
- * cancels it. Matches the gesture used by Finder/Notes for "click =
- * open, double-click = rename".
- *
- * Commit semantics:
- *   - Trim → compare to original. If unchanged: silent revert.
- *   - If changed and ≤ 300 chars: call onCommit. Parent owns the
- *     PATCH + optimistic update + rollback on failure.
- *   - If empty after trim: silent revert (no destructive rename).
- *
- * Doesn't bubble Enter / clicks while editing — those would
- * otherwise toggle the parent row's complete state or open
- * comments. `stopPropagation` on the key + click handlers
- * isolates the input.
- */
-function InlineTitleEditor({
-  title,
-  done,
-  href,
-  onCommit,
-}: {
-  title: string;
-  done: boolean;
-  /**
-   * Destination for a single-click on the title (the task detail page).
-   * The row's own click handler still selects the row; navigation runs
-   * on its own short timer so a dblclick can cancel it and enter
-   * rename mode instead.
-   */
-  href: string;
-  onCommit: (next: string) => void;
-}) {
-  const router = useRouter();
-  const [editing, setEditing] = useState(false);
-  const [draft, setDraft] = useState(title);
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  // Holds the pending navigation timer started by a single click. A
-  // subsequent dblclick clears it so we don't navigate-and-rename in
-  // the same gesture.
-  const clickTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  // Reset the draft whenever the upstream title changes — handles the
-  // case where another collaborator renames the row while we're not
-  // editing (realtime push lands).
-  useEffect(() => {
-    if (!editing) setDraft(title);
-  }, [title, editing]);
-
-  // Auto-select on focus so a double-click → type-to-replace flow
-  // feels native (matches Finder rename, GitHub issue titles, etc.).
-  useEffect(() => {
-    if (editing) {
-      inputRef.current?.focus();
-      inputRef.current?.select();
-    }
-  }, [editing]);
-
-  // Cancel any pending nav timer on unmount so a row that scrolls off
-  // mid-gesture doesn't try to navigate later.
-  useEffect(() => {
-    return () => {
-      if (clickTimerRef.current) clearTimeout(clickTimerRef.current);
-    };
-  }, []);
-
-  function commit() {
-    const next = draft.trim();
-    setEditing(false);
-    if (!next || next === title) {
-      // Empty or unchanged — silent revert.
-      setDraft(title);
-      return;
-    }
-    onCommit(next);
-  }
-
-  function cancel() {
-    setDraft(title);
-    setEditing(false);
-  }
-
-  if (!editing) {
-    return (
-      <span
-        className={cn(
-          "cursor-pointer truncate text-sm hover:underline",
-          done ? "text-text-3 line-through" : "text-text-0",
-        )}
-        role="link"
-        title="Click to open · double-click to rename"
-        onClick={(e) => {
-          // Don't bubble — the row's own onClick should still fire to
-          // select the row, but we want full control over the timer.
-          // `setSelectedIdx` lives on the outer row click; calling it
-          // here too is redundant but harmless. We stopPropagation to
-          // avoid double-selection animations.
-          e.stopPropagation();
-          if (clickTimerRef.current) clearTimeout(clickTimerRef.current);
-          clickTimerRef.current = setTimeout(() => {
-            clickTimerRef.current = null;
-            router.push(href);
-          }, 220);
-        }}
-        onDoubleClick={(e) => {
-          e.stopPropagation();
-          // Cancel the pending navigate from the preceding single click.
-          if (clickTimerRef.current) {
-            clearTimeout(clickTimerRef.current);
-            clickTimerRef.current = null;
-          }
-          setEditing(true);
-        }}
-      >
-        {title}
-      </span>
-    );
-  }
-
-  return (
-    <input
-      ref={inputRef}
-      type="text"
-      value={draft}
-      onChange={(e) => setDraft(e.target.value)}
-      onClick={(e) => e.stopPropagation()}
-      onMouseDown={(e) => e.stopPropagation()}
-      onKeyDown={(e) => {
-        e.stopPropagation();
-        if (e.key === "Enter") {
-          e.preventDefault();
-          commit();
-        } else if (e.key === "Escape") {
-          e.preventDefault();
-          cancel();
-        }
-      }}
-      onBlur={commit}
-      maxLength={300}
-      aria-label="Task title"
-      className={cn(
-        "min-w-0 flex-1 truncate rounded-sm border border-border-focus bg-bg-0 px-1 py-0.5 text-sm text-text-0 outline-none",
-        done && "line-through",
-      )}
-    />
   );
 }
