+++ apps/web/src/components/dashboard/TasksCard.tsx
@@ -1,17 +1,10 @@
 "use client";
 
-import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { Check } from "lucide-react";
-import { cn } from "@/lib/utils";
-import { PriorityDots, DueChip } from "@/components/primitives";
-import {
-  TaskSectionHeader,
-  groupTone,
-  priorityEdge,
-} from "@/components/TaskSectionHeader";
+import { TaskSectionHeader, groupTone } from "@/components/TaskSectionHeader";
 import { TaskListToolbar, type GroupOption } from "@/components/TaskListToolbar";
+import { TaskRow } from "@/components/TaskRow";
 import { useRealtimeTable } from "@/lib/supabase/realtime";
 import { bucketDashTasks } from "@/lib/task-grouping";
 import type { AssignedTask, DelegatedTask } from "@/lib/dashboard-queries";
@@ -296,7 +289,7 @@ export function TasksCard({
         ) : groupBy === "none" ? (
           <ul className="divide-y divide-border">
             {visibleAssigned.slice(0, ROW_LIMIT).map((t) => (
-              <AssignedRow
+              <DashboardTaskRow
                 key={t.id}
                 task={t}
                 ticker={tickerById[t.terminal_id]}
@@ -331,7 +324,7 @@ export function TasksCard({
                       {!collapsed ? (
                         <ul className="divide-y divide-border">
                           {b.tasks.map((t) => (
-                            <AssignedRow
+                            <DashboardTaskRow
                               key={`${b.key}:${t.id}`}
                               task={t}
                               ticker={tickerById[t.terminal_id]}
@@ -352,7 +345,15 @@ export function TasksCard({
   );
 }
 
-function AssignedRow({
+/**
+ * A dashboard task row. Thin wrapper that owns optimistic done/star
+ * state and delegates ALL rendering to the shared {@link TaskRow}, so
+ * the dashboard row is pixel-identical to the in-terminal pane —
+ * square checkbox, amber star, priority left-edge + dots, status pill.
+ * Toggling done or star PATCHes /api/v1/tasks/:id (same endpoint the
+ * terminal uses) and reconciles on the next realtime-driven refresh.
+ */
+function DashboardTaskRow({
   task,
   ticker,
   terminalName,
@@ -361,112 +362,50 @@ function AssignedRow({
   ticker?: string;
   terminalName?: string;
 }) {
-  const [optimisticDone, setOptimisticDone] = useState<boolean | null>(null);
-  const isDone =
-    optimisticDone !== null ? optimisticDone : task.status === "done";
-  // Deep-link to the task detail surface so a click puts the user one step
-  // away from the work, not just on the parent terminal page.
-  const href = ticker ? `/p/${ticker}/task/${task.ticker_seq}` : undefined;
+  // Local optimistic overrides for the two togglable fields. Applied on
+  // top of the server-provided task; cleared (rolled back) if the PATCH
+  // fails. The dashboard refetches via realtime, which supersedes these.
+  const [override, setOverride] = useState<
+    Partial<Pick<AssignedTask, "status" | "starred">>
+  >({});
+  const merged: AssignedTask = { ...task, ...override };
 
-  /**
-   * Toggle done directly from the dashboard row. The status icon
-   * used to be decorative (just a Check / Circle); it's now a
-   * button so the user can flip a task without navigating into
-   * the detail page. Optimistic — flips locally on click,
-   * reconciles silently against the server response.
-   */
-  async function toggleDone(e: React.MouseEvent) {
-    e.preventDefault();
-    e.stopPropagation();
-    const next = !isDone;
-    setOptimisticDone(next);
+  function rollback(patch: Partial<Pick<AssignedTask, "status" | "starred">>) {
+    setOverride((o) => {
+      const next = { ...o };
+      for (const k of Object.keys(patch)) delete next[k as keyof typeof next];
+      return next;
+    });
+  }
+
+  async function commit(
+    patch: Partial<Pick<AssignedTask, "status" | "starred">>,
+  ) {
+    setOverride((o) => ({ ...o, ...patch }));
     try {
       const r = await fetch(`/api/v1/tasks/${task.id}`, {
         method: "PATCH",
         credentials: "include",
         headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ status: next ? "done" : "todo" }),
+        body: JSON.stringify(patch),
       });
-      if (!r.ok) setOptimisticDone(null);
+      if (!r.ok && r.status !== 202) rollback(patch);
     } catch {
-      setOptimisticDone(null);
+      rollback(patch);
     }
   }
 
-  // Two adjacent click targets in one row:
-  //   - the circle button → toggle done (no nav)
-  //   - the rest of the row → navigate to task detail
-  //
-  // Critically, the button must NOT be nested inside the Link.
-  // `<button>` inside `<a>` is invalid HTML and browsers handle the
-  // clash inconsistently — Zack reported "I can't complete tasks
-  // from the dashboard, I have to open the detail." Separating
-  // them as siblings is the reliable fix; stopPropagation on the
-  // button was already in place but couldn't help once the Link
-  // navigation was queued.
-  //
-  // The row gets the `hover:bg-bg-2` so hovering anywhere in the
-  // row still feels like one unit. Both children are flex items
-  // inside the same `<li>`.
-  const linkContent = (
-    <>
-      <span
-        className={cn(
-          "flex-1 truncate",
-          isDone ? "text-text-3 line-through" : "text-text-0",
-        )}
-      >
-        {task.title}
-      </span>
-      {terminalName ? (
-        <span
-          className="hidden truncate text-text-3 md:inline max-w-[10ch]"
-          title={terminalName}
-        >
-          {terminalName}
-        </span>
-      ) : null}
-      <PriorityDots priority={task.priority} />
-      {task.due_date ? <DueChip date={task.due_date} /> : null}
-    </>
-  );
-
   return (
-    <li
-      data-row
-      className={cn(
-        // 2px priority left-edge (High red, Medium amber) matching the
-        // in-terminal pane; pl-[10px] + 2px border = the base px-3 so
-        // colouring never shifts the row.
-        "flex items-center gap-2 border-l-2 px-3 py-1 pl-[10px] text-xs hover:bg-bg-2",
-        priorityEdge(task.priority),
-      )}
-    >
-      <button
-        type="button"
-        onClick={toggleDone}
-        aria-label={isDone ? "Mark as not done" : "Mark as done"}
-        className={cn(
-          "flex h-3.5 w-3.5 flex-shrink-0 items-center justify-center rounded-full border transition-colors",
-          isDone
-            ? "border-success bg-success-subtle text-success"
-            : "border-text-3 hover:border-accent",
-        )}
-      >
-        {isDone ? <Check className="h-2.5 w-2.5" aria-hidden="true" /> : null}
-      </button>
-      {href ? (
-        <Link
-          href={href}
-          className="flex flex-1 items-center gap-2 min-w-0"
-        >
-          {linkContent}
-        </Link>
-      ) : (
-        <div className="flex flex-1 items-center gap-2 min-w-0">
-          {linkContent}
-        </div>
-      )}
+    <li>
+      <TaskRow
+        task={merged}
+        ticker={ticker ?? ""}
+        terminalName={terminalName}
+        onToggle={() =>
+          commit({ status: merged.status === "done" ? "todo" : "done" })
+        }
+        onToggleStar={() => commit({ starred: !merged.starred })}
+      />
     </li>
   );
 }
